requirements.txt
# LangChain and ChromaDB dependencies
langchain
chromadb  # No need for 'langchain-chromadb', since Chroma is built-in
langchain-community
langchain-openai
langchain-xai
# Additional libraries
datasets
pypdf
python-dotenv
pandas
flask

# Editable install for local package
-e .